fix carpool /myCarpool, add latitude, longitude + not yet done S3

+++ src/main/java/MATE/Carpool/domain/S3/service/AwsS3Service.java
@@ -1,23 +1,19 @@
 package MATE.Carpool.domain.S3.service;
 
+import MATE.Carpool.common.Message;
 import MATE.Carpool.common.exception.CustomException;
 import MATE.Carpool.common.exception.ErrorCode;
-import MATE.Carpool.config.userDetails.CustomUserDetails;
-import MATE.Carpool.domain.member.entity.Member;
 import MATE.Carpool.domain.member.repository.MemberRepository;
-import MATE.Carpool.domain.member.service.MemberService;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Random;
 import java.util.UUID;
 
 @Service
@@ -50,16 +46,22 @@ public class AwsS3Service {
 
     // S3 파일 삭제 메서드
     public ResponseEntity<String> deleteImg(String imgKey) {
+
         try {
-            // 해당 키에 대한 파일이 존재하는지 확인
-            if (s3Client.doesObjectExist(bucketName, imgKey)) {
-                // 파일 삭제
-                s3Client.deleteObject(bucketName, imgKey);
-                return ResponseEntity.ok("이미지가 삭제되었습니다: " + imgKey);
-                //System.out.println("이미지가 삭제되었습니다: " + imgKey);
-            } else {
-                throw new CustomException(ErrorCode.RESOURCE_NOT_FOUND);
+            //기본 이미지 체크
+            if (imgKey.equals("profileImgS3.png")) {
+                return ResponseEntity.ok("기본 이미지는 삭제할 수 없습니다.");
             }
+
+                // 해당 키에 대한 파일이 존재하는지 확인
+                if (s3Client.doesObjectExist(bucketName, imgKey)) {
+                    // 파일 삭제
+                    s3Client.deleteObject(bucketName, imgKey);
+                    return ResponseEntity.ok("이미지가 삭제되었습니다: " + imgKey);
+                    //System.out.println("이미지가 삭제되었습니다: " + imgKey);
+                } else {
+                    throw new CustomException(ErrorCode.RESOURCE_NOT_FOUND);
+                }
         } catch (Exception e) {
             throw new RuntimeException("이미지 삭제 중 오류가 발생했습니다.", e);
         }
